requirements.txt
flask~=3.1.0
numpy~=2.0.2
pillow~=11.1.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
tf_keras~=2.18.0
